Ajout des composants CartActions, CartSummary et CarItem pour améliorer la gestion du panier, ainsi que la refactorisation du composant Cart pour intégrer ces nouvelles fonctionnalités.
